# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tabi API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/items/{item_id}/checkins": {
      "post": {
        "tags": [
          "checkins"
        ],
        "summary": "Create Checkin",
        "description": "Create one experience and return the original row on a matching retry.",
        "operationId": "create_checkin_api_items__item_id__checkins_post",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 80,
              "title": "Idempotency-Key"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/checkins": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "History",
        "description": "Read own active records, including archived checklist entries.",
        "operationId": "history_api_me_checkins_get",
        "parameters": [
          {
            "name": "list_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "List Id"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OwnCheckinOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkins/{checkin_id}": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "Own Detail",
        "description": "Read one owned experience.",
        "operationId": "own_detail_api_checkins__checkin_id__get",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "checkins"
        ],
        "summary": "Edit Checkin",
        "description": "Edit one owned experience without affecting other visits.",
        "operationId": "edit_checkin_api_checkins__checkin_id__patch",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckinEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "checkins"
        ],
        "summary": "Delete Checkin",
        "description": "Soft-delete one record and immediately remove it from progress/shares.",
        "operationId": "delete_checkin_api_checkins__checkin_id__delete",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/items/{item_id}/checkins/public": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "Public Experiences",
        "description": "Page through visible experiences on a published item.",
        "operationId": "public_experiences_api_items__item_id__checkins_public_get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_PublicCheckinOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/shares/{share_id}": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "Shared Experience",
        "description": "Resolve a stable share ID only while the record remains public.",
        "operationId": "shared_experience_api_shares__share_id__get",
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCheckinOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/checkins/{checkin_id}/hide": {
      "post": {
        "tags": [
          "checkins"
        ],
        "summary": "Hide Experience",
        "description": "Hide a public record and record why it was moderated.",
        "operationId": "hide_experience_api_admin_checkins__checkin_id__hide_post",
        "parameters": [
          {
            "name": "checkin_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Checkin Id"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HideIn"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/audit": {
      "get": {
        "tags": [
          "checkins"
        ],
        "summary": "Audit History",
        "description": "Read content audit entries as a system administrator.",
        "operationId": "audit_history_api_admin_audit_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinCreate": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "experienced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experienced At"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public"
            ],
            "title": "Visibility",
            "default": "private"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "upload_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "title": "Upload Ids"
          }
        },
        "type": "object",
        "title": "CheckinCreate",
        "description": "One new experience with optional staged images."
      },
      "CheckinEdit": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "experienced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experienced At"
          },
          "visibility": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "private",
                  "public"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Visibility"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionIn"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "CheckinEdit",
        "description": "Fields editable on one existing experience."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HideIn": {
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "HideIn",
        "description": "Administrator moderation reason."
      },
...
      "OwnCheckinOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "share_id": {
            "type": "string",
            "title": "Share Id"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "item_name": {
            "type": "string",
            "title": "Item Name"
          },
          "list_id": {
            "type": "string",
            "title": "List Id"
          },
          "list_title": {
            "type": "string",
            "title": "List Title"
          },
          "author_id": {
            "type": "string",
            "title": "Author Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "author_avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Author Avatar Url"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "experienced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Experienced At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "media": {
            "items": {
              "$ref": "#/components/schemas/MediaOut"
            },
            "type": "array",
            "title": "Media"
          },
          "visibility": {
            "type": "string",
            "enum": [
              "private",
              "public"
            ],
            "title": "Visibility"
          },
          "hidden": {
            "type": "boolean",
            "title": "Hidden"
          },
          "position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PositionIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "item_status": {
            "type": "string",
            "title": "Item Status"
          },
          "list_status": {
            "type": "string",
            "title": "List Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "share_id",
          "item_id",
          "item_name",
          "list_id",
          "list_title",
          "author_id",
          "author_name",
          "author_avatar_url",
          "note",
          "experienced_at",
          "created_at",
          "media",
          "visibility",
          "hidden",
          "position",
          "updated_at",
          "item_status",
          "list_status"
        ],
        "title": "OwnCheckinOut",
        "description": "Owner view including private state and optional position."
      },
...
      "Page_OwnCheckinOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnCheckinOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[OwnCheckinOut]"
      },
      "Page_PublicCheckinOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicCheckinOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[PublicCheckinOut]"
      },
      "Page_dict_": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[dict]"
      },
...
    }
  }
}